Refactor name to prisma-mock
