+++ jest.config.js
@@ -2,7 +2,7 @@ module.exports = {
     clearMocks: true,
     preset: 'ts-jest',
     testEnvironment: 'node',
-    setupFilesAfterEnv: ['<rootDir>/src/singleton.ts'],
+    setupFilesAfterEnv: ['<rootDir>/src/prisma-mock.ts'],
     moduleNameMapper: {
         '^@/(.*)$': '<rootDir>/src/$1',
     },
