+++ src/prisma-mock.ts
@@ -1,8 +1,8 @@
 import { PrismaClient } from '@prisma/client'
 import { mockDeep, mockReset, DeepMockProxy } from 'jest-mock-extended'
-import prisma from './lib/prisma'
+import prisma from '@/lib/prisma'
 
-jest.mock('./lib/prisma', () => ({
+jest.mock('@/lib/prisma', () => ({
     __esModule: true,
     default: mockDeep<PrismaClient>(),
 }))
